refactor: split loop.py into focused modules using mixins
Extract the monolithic loop.py (~1100 lines) into responsibility-based
modules while preserving the identical public API:

- exceptions.py: exception hierarchy (PyNomalyError, ValidationError, etc.)
- _utils.py: Utils class (progress bar)
- _validation.py: ValidationMixin + @accepts decorator
- _distance.py: DistanceMixin + Numba JIT kernels
- _pipeline.py: PipelineMixin (LoOP scoring math)
- loop.py: slim orchestrator composing the class from mixins

No API changes. All existing import patterns continue to work.
